Fixed dependency cycle
