+++ autoscheduler/frontend/src/redux/reducers/term.ts
@@ -2,16 +2,10 @@
  * Stores the current term
  */
 
-import { TermDataAction } from './termData';
+import { TermDataAction } from '../actions/termData';
 
 // action type string
 export const SET_TERM = 'SET_TERM';
-
-// action type interface
-export interface SetTermAction {
-  type: 'SET_TERM';
-  term: string;
-}
 
 // reducer
 export default function term(state = '', action: TermDataAction): string {
